# ms.teachers.management/src/main/java/com/example/ms/teachers/management/core/port/in/dto/SalaryDto.java
package com.example.ms.teachers.management.core.port.in.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Data;

@Data
public class SalaryDto   {
  @JsonProperty("amount")
  private Double amount = null;

  @JsonProperty("currency")
  private String currency = null;

}